requirements.txt
backports.functools-lru-cache==1.5
certifi==2019.3.9
chardet==3.0.4
cheroot==6.5.5
CherryPy==17.4.1
contextlib2==0.5.5
idna==2.8
jaraco.functools==2.0
Jinja2==2.10.1
MarkupSafe==1.1.1
more-itertools==5.0.0
portend==2.4
pytz==2019.1
PyYAML==5.1
redis==3.2.1
requests==2.22.0
six==1.12.0
tempora==1.14.1
urllib3==1.25.3
zc.lockfile==1.4
